requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.8
discord.py==2.4.0
dnspython==2.1.0
ffmpeg-python==0.2.0
ffprobe-python==1.0.3
Pillow
webcolors
psutil
colorlog
asyncpg
spotipy
isodate
future==0.18.2
idna==3.3
multidict==5.2.0
mutagen==1.46.0
pycparser==2.21
pycryptodomex==3.17
pymongo==3.12.2
PyNaCl==1.4.0
requests==2.26.0
six==1.16.0
typing_extensions==4.0.0
urllib3==1.26.7
websockets==11.0.1
yarl==1.7.2
yt-dlp
